MoE a16wmix: move the duplicated bf16 MMA helper into utils
_bf16_frag / _bf16_frag4 / _mma were byte-identical in gemm1.py and gemm2.py
(21 lines duplicated). Moved to utils.mma_bf16 (+ two private frag builders);
each kernel now binds it with functools.partial(mma_bf16, mma_atom, use_k16),
so the mainloop call sites are unchanged.

gemm1 1159->1141, gemm2 820->802, utils 179->204: net -13 lines and one copy
of the MMA fragment logic instead of two. Codegen-neutral: mxfp4 gemm1+gemm2
ISA byte-identical to origin/main, a16wfp4 pytest 12/12, int4 logits_diff
1.41-1.57e-5 at tok 1/4/16384.
